requirements.txt
# UniRec - Python Dependencies

# Core ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Recommendation Libraries
lightgbm>=4.0.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic[email]>=2.0.0

# Authentication
pyjwt>=2.8.0
bcrypt>=4.0.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.1.0

# Development
pytest>=7.4.0
requests>=2.31.0